made gtt use gnome-stock.[ch]
Mon Jan 19 15:52:46 1998 MET  Eckehard Berns  <eb@berns.prima.de>

	* programs/gtt: made gtt use gnome-stock.[ch]

	* libgnomeui/libgnomeui.h: include gnome-stock.h

	* libgnomeui/gnome-stock.[ch]: added. They provide stock icons for
 	  toolbars etc. Short description in gnome-stock.h. Not everything
 	  is implemented, espacially theme support. But they're usable.

+++ gtt/proj.c
@@ -408,7 +408,7 @@ char *project_get_timestr(project *proj)
 	} else {
 		t = proj->day_secs;
 	}
-	sprintf(s, "%02d:%02d", (t / 3600), (t % 3600) / 60);
+	sprintf(s, "%02d:%02d", (int)(t / 3600), (int)((t % 3600) / 60));
 	return s;
 }
